rename hardwallet to keycard
