+++ README.md
@@ -1,8 +1,8 @@
-# Hardwallet Android SDK
+# Keycard Android SDK
 
-This SDK simplifies integration with the [Status Keycard](https://github.com/status-im/hardware-wallet) in Android
+This SDK simplifies integration with the [Status Keycard](https://github.com/status-im/status-keycard) in Android
 applications. In this SDK you find both the classes needed for generic communication with SmartCards as well as classes
 specifically addressing the Keycard.
 
-To get started, check the file demo/src/main/java/im/status/hardwallet_lite_android/app/MainActivity.java which a simple
+To get started, check the file demo/src/main/java/im/status/keycard/app/MainActivity.java which a simple
 demo application showing how the SDK works and what you can do with it.
