+++ demo/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "im.status.hardwallet_lite_android.demo"
+        applicationId "im.status.keycard.demo"
         minSdkVersion 19
         targetSdkVersion 28
         versionCode 1
