+++ demo/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="im.status.hardwallet_lite_android.demo">
+          package="im.status.keycard.demo">
 
     <application
             android:allowBackup="true"
@@ -9,7 +9,7 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name="im.status.hardwallet_lite_android.app.MainActivity">
+        <activity android:name="im.status.keycard.app.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
