+++ demo/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="im.status.hardwallet_lite_android.app.MainActivity">
+        tools:context="im.status.keycard.app.MainActivity">
 
     <TextView
             android:layout_width="wrap_content"
