+++ demo/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">HardwalletLiteAndroid</string>
+    <string name="app_name">KeycardAndroid</string>
 </resources>
