--- lib/src/main/java/im/status/hardwallet_lite_android/io/CardListener.java
@@ -1,6 +1,0 @@
-package im.status.hardwallet_lite_android.io;
-
-public interface CardListener {
-    void onConnected(CardChannel channel);
-    void onDisconnected();
-}
