+++ lib/src/main/java/im/status/keycard/io/CardListener.java
@@ -1,0 +1,6 @@
+package im.status.keycard.io;
+
+public interface CardListener {
+    void onConnected(CardChannel channel);
+    void onDisconnected();
+}
